Add Prettier script and config
This changes spacing from 4 to 2. The actual code is near identical other than line lengths being a bit shorter in JS (could be modified in the prettier config if you want?).
For the CSS a leading zero was adding to `filter .1s` (line 58) and one instance of double-quotes was changed to single quotes (line 31).